refactor(utils): move logsumexp_normalize to treetime-utils as softmax_with_log_norm
- Move function from marginal_helpers.rs to treetime-utils/src/array/softmax_with_log_norm.rs
- Rename to reflect semantics: returns (softmax, log_norm) tuple
- Add KaTeX math notation in doc comment explaining softmax, LSE, numerical stability, and fusion
- Update all call sites in discrete.rs, marginal_dense.rs, marginal_helpers.rs, gtr_refinement.rs, infer_gtr/dense.rs
- Update doc references in port-algo-inventory, port-test-inventory, port-known-issues, and reports
