+++ docs/port-algo-inventory/ancestral.md
@@ -65,7 +65,7 @@ v0: [`packages/legacy/treetime/treetime/treeanc.py#L762-L927`](../../packages/le
 
 ### v0 differences
 
-v1 backward pass uses log-space arithmetic with logsumexp normalization (dense `normalize_from_log()`, sparse `logsumexp_normalize()` in `combine_messages()`). v1 forward pass uses plain probability space (division). v0 uses neg-log space throughout. v1 dense uses deterministic `argmax_first()` (`#argmax_first`) for sequence extraction (leftmost state wins ties); v0 uses `np.argmax()` which has undefined tie-breaking.
+v1 backward pass uses log-space arithmetic with logsumexp normalization (dense `normalize_from_log()`, sparse `softmax_with_log_norm()` in `combine_messages()`). v1 forward pass uses plain probability space (division). v0 uses neg-log space throughout. v1 dense uses deterministic `argmax_first()` (`#argmax_first`) for sequence extraction (leftmost state wins ties); v0 uses `np.argmax()` which has undefined tie-breaking.
 
 ### Key functions
