+++ docs/port-test-inventory/ancestral.md
@@ -302,25 +302,20 @@ Support files (helpers only, no tests): `prop_marginal_support.rs`, `test_margin
 
 ---
 
-## Logsumexp Normalize Tests
+## Softmax with Log-Norm Tests
 
-**File:** [`marginal_helpers.rs`](../../packages/treetime/src/representation/partition/marginal_helpers.rs) (inline `#[cfg(test)]`)
+**File:** [`softmax_with_log_norm.rs`](../../packages/treetime-utils/src/array/softmax_with_log_norm.rs) (inline `#[cfg(test)]`)
 
-| Test                                                  | Purpose                                                  |
-| ----------------------------------------------------- | -------------------------------------------------------- |
-| `test_logsumexp_normalize_equal_log_probs`            | Equal log-probs produce uniform distribution, log_norm=0 |
-| `test_logsumexp_normalize_descending`                 | Unequal log-probs: analytically verified output and norm |
-| `test_logsumexp_normalize_single_dominant_state`      | One high state concentrates all probability mass         |
-| `test_logsumexp_normalize_all_neg_inf`                | All-NEG_INFINITY fallback: uniform 1/4, log_norm=-inf    |
-| `test_logsumexp_normalize_all_neg_inf_three_states`   | All-NEG_INFINITY fallback with 3 states: uniform 1/3     |
-| `test_logsumexp_normalize_mixed_finite_neg_inf`       | Mixed finite/-inf: only finite states get probability    |
-| `test_logsumexp_normalize_single_finite_rest_neg_inf` | One finite state among -inf: gets all mass               |
-| `test_logsumexp_normalize_large_negative_values`      | Numerical stability: offset -1000 matches reference      |
-| `test_logsumexp_normalize_large_positive_values`      | Numerical stability: offset +1003 matches reference      |
-| `test_logsumexp_normalize_single_state`               | Single element: probability 1.0, log_norm=input value    |
-| `test_logsumexp_normalize_two_equal_states`           | Two equal values: [0.5, 0.5], log_norm=ln(2)             |
+| Test                                         | Purpose                                                |
+| -------------------------------------------- | ------------------------------------------------------ |
+| `test_softmax_with_log_norm_finite`          | All-finite inputs: softmax has no zeros, all positive  |
+| `test_softmax_with_log_norm_with_neg_inf`    | Mixed finite/-inf: only finite states get probability  |
+| `test_softmax_with_log_norm_degenerate`      | All-NEG_INFINITY fallback: uniform, log_norm=-inf      |
+| `test_softmax_with_log_norm_uniform_input`   | Equal finite inputs: exact uniform output              |
+| `test_softmax_with_log_norm_shift_invariant` | Shift invariance: constant offset preserves softmax    |
+| `stress::test_softmax_with_log_norm_*`       | Numerical stability: overflow/underflow boundary tests |
 
-**Algorithm:** Logsumexp trick for numerically stable softmax normalization. Subtracts max before exponentiation to prevent overflow/underflow.
+**Algorithm:** Fused logsumexp + softmax for numerically stable normalization. Subtracts max before exponentiation to prevent overflow/underflow. Returns both softmax result and log normalization constant.
 
 ---
