+++ packages/treetime/src/gtr/infer_gtr/dense.rs
@@ -1,6 +1,8 @@
 use crate::constants::SUPERTINY_NUMBER;
 use crate::gtr::gtr::{GTR, GTRParams};
-use crate::gtr::infer_gtr::common::{InferGtrOptions, InferGtrResult, MutationCounts, infer_gtr_impl, is_profile_informative};
+use crate::gtr::infer_gtr::common::{
+  InferGtrOptions, InferGtrResult, MutationCounts, infer_gtr_impl, is_profile_informative,
+};
 use crate::hacks::fix_branch_length::fix_branch_length;
 use crate::make_internal_report;
 use crate::representation::partition::marginal_dense::PartitionMarginalDense;
